test: cmd/agent тесты счетчиков вынесены в отдельный файл
